Plot Method for simmer Objects

Deprecated. See plot.mon instead.

Usage,

# S3 method for simmer
plot(x, what = c("resources", "arrivals", "attributes"), metric = NULL, ...)

Arguments

x

a single simmer environment or a list of environments representing several replications.

what

type of plot, one of c("resources", "arrivals", "attributes").

metric

specific metric for each type of plot.

what = "resources"

one of c("usage", "utilization").

what = "arrivals"

one of c("activity_time", "waiting_time", "flow_time").

what = "attributes"

no metrics at the moment.

...

further arguments for each kind of plot.

what = "resources"

all metrics

names

the name of the resource(s) (a single string or a character vector) to show.

metric = "usage"

items

the components of the resource to be plotted, one or more of c("system", "queue", "server").

steps

if TRUE, shows the instantaneous usage instead of the cumulative average.

what = "attributes"

keys

the keys of attributes you want to plot (if left empty, all attributes are shown).

Value

Returns a ggplot2 object.